+++ board-frontend/src/scripts/main.js
@@ -1,1 +1,2 @@
 import './Document.js';
+import './FileUpload.js'