+++ CHANGELOG.md
@@ -129,3 +129,11 @@ Proper cookie authentication
 ### v0.4.4
 Bugfixes, better urls/file paths
 - Fix cookies not having proper security properties
+- Improve routing
+
+## v0.5
+Canvas features
+
+### v0.5.0
+Image upload
+- Ability to drag and drop images on to the board